# reports/openapi_day4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCM Copilot",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/kb/chat": {
      "post": {
        "tags": [
          "kb"
        ],
        "summary": "知识问答（SSE 流式）",
        "description": "混合检索（BM25+向量+RRF+重排）→ 生成+双校验（CRAG/缺失回退）→ 流式回答 + 引用溯源。事件协议见下方 200 响应描述（progress/message/citations/data_table/done/error）。需要权限 kb:chat。",
        "operationId": "chat_api_v1_kb_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KbChatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "SSE 流（text/event-stream），事件协议：\n- progress:  {type, node, data:{result}}——链路节点进展\n- message:   {type, role, content, delta, session_id}——打字机增量\n- citations: {type, citations, retrieved_docs, validation, source?, session_id}——引用溯源\n- data_table:{type, columns, rows, sql, insights, ...}——★ W24 Day6 查数表格事件\n- done:      {type}——流结束；error: {type, error}——链路异常",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"type\":\"progress\",\"node\":\"retrieve\",\"data\":{\"result\":\"混合检索命中 3 个候选\"}}\n\n"
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/kb/feedback": {
      "post": {
        "tags": [
          "kb"
        ],
        "summary": "引用纠错反馈",
        "description": "点赞/纠错 → 待审核 → 管理员审核 → 回流评测集 v2。需要权限 kb:feedback。",
        "operationId": "feedback_api_v1_kb_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KbFeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KbFeedbackOut"
                }
              }
            }
          },
          "401": {
            "description": "未认证 / 凭证无效 / 过期 / 已吊销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "403": {
            "description": "已认证但权限码未命中",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "429": {
            "description": "限流 / API Key 配额超限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "500": {
            "description": "服务内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          },
          "503": {
            "description": "依赖服务不可用（如调度器）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Err": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "机器可读错误码，如 AUTH_401 / QUOTA_429"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "人类可读错误说明"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "description": "贯穿一次请求的 request_id（排查定位用）"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "Err",
        "description": "统一错误响应体（OpenAPI `responses={401: {\"model\": Err}, ...}` 引用）。"
      },
...
      "KbChatIn": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "用户问题（服务端限制 ≤2000 字符）"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话标识；缺省服务端生成"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "KbChatIn",
        "description": "知识问答请求体（SSE 流式）。\n\n长度限制由服务端显式校验（空消息/超 2000 字符 → 400），不在模型层强制\n（保持与既有行为一致；模型层只做结构契约）。"
      },
      "KbFeedbackIn": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question",
            "description": "原始问题"
          },
          "action": {
            "type": "string",
            "enum": [
              "like",
              "dislike",
              "correct"
            ],
            "title": "Action",
            "description": "反馈动作",
            "default": "like"
          },
          "original_answer": {
            "type": "string",
            "title": "Original Answer",
            "description": "原回答",
            "default": ""
          },
          "corrected_answer": {
            "type": "string",
            "title": "Corrected Answer",
            "description": "纠正后的回答",
            "default": ""
          },
          "correct_doc_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Correct Doc Ids",
            "description": "正确引用文档 id"
          },
          "qa_id": {
            "type": "string",
            "title": "Qa Id",
            "description": "关联问答 id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "KbFeedbackIn",
        "description": "引用纠错反馈请求体（点赞/纠错 → 待审核 → 回流评测集 v2）。"
      },
      "KbFeedbackOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "feedback_id": {
            "type": "string",
            "title": "Feedback Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "feedback_id",
          "status"
        ],
        "title": "KbFeedbackOut",
        "description": "反馈提交响应。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}